✨ 기능 수정: UserPerCampaignRepository -> Id 타입 Integer로 변경

+++ LearnsMate/src/main/java/intbyte4/learnsmate/userpercampaign/repository/UserPerCampaignRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UserPerCampaignRepository extends JpaRepository<UserPerCampaign, Long> {
+public interface UserPerCampaignRepository extends JpaRepository<UserPerCampaign, Integer> {
     List<UserPerCampaign> findByCampaign(Campaign campaign);
 }
